+++ packages/assistant/tests/python/assistant_test.py
@@ -54,14 +54,16 @@ assistant.add_icon(Icon.Failure)
 assistant.add_icon(Icon.Warning)
 assistant.add_icon(Icon.Success)
 assistant.add_text_input(
-    "txt_input", placeholder="placeholder", validation=length_greater_3
+    "txt_input", placeholder="placeholder", required=True, validation=length_greater_3
 )
 assistant.add_password_input("pw_input")
 assistant.add_checkbox("checkbox", "test_checkbox")
 assistant.add_file_input("file", file_type="pdf,png,jpg")
 assistant.add_hidden_input("Hidden", "value")
 assistant.add_text("Percentage slider")
-assistant.add_slider(name="percentage", slider_min=0, slider_max=0, steps=100)
+assistant.add_slider(
+    name="percentage", slider_min=0, slider_max=0, steps=100, default=80
+)
 # assistant.add_file(path="/Users/kerkko/Downloads/image.png", label="File")
 
 assistant.add_radio_buttons(